refactor: simplify markdown convert
